requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
docx2pdf==0.1.8
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
python-dotenv==1.0.0
pywin32==306
tqdm==4.66.1
Werkzeug==3.0.1
